MINOR: Move `ControllerPurgatory` to `server-common`
